+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ### webapp
 
 - feat: added unit conversion tool for Length, Temperature, Time, Data Size
+- feat: favourite and pin tools in webapp
 
 ## [v1.2.0] - 2025-12-30